Offload through dedicated instructions works in complex core. Conventional HWPE configuration does not build the hardware due to strange QUESTA internal error linked to CV32E40P CSR endmodule. Need to investigate this and to adapt new reduced configuration based on internal tiler to the conventional memory-mapped approach.
