+++ Makefile
@@ -25,13 +25,14 @@ QUESTA         ?= questa-2022.3
 QUESTA_HOME    ?= /usr/pack/modelsim-10.7b-kgf/questasim
 BENDER_DIR     ?= .
 BENDER         ?= bender
-TEST_SRCS      ?= sw/redmule.c
-WAVES          ?= $(mkfile_path)/wave.do
 ISA            ?= riscv
 ARCH           ?= rv
-XLEN           ?= 32
-# XTEN           ?= imc
-XTEN           ?= im
+
+ifeq ($(REDMULE_COMPLEX),1)
+	TEST_SRCS := sw/redmule_complex.c
+else
+	TEST_SRCS := sw/redmule.c
+endif
 
 compile_script ?= scripts/compile.tcl
 compile_flag   ?= -suppress 2583 -suppress 13314
@@ -45,15 +46,22 @@ ipstools ?= 0
 P_STALL  ?= 0.0
 
 ifeq ($(verbose),1)
-FLAGS += -DVERBOSE
+	FLAGS += -DVERBOSE
 endif
 
-# Setup toolchain (from SDK) and options
-CC=$(PULP_RISCV_GCC_TOOLCHAIN)/bin/$(ISA)$(XLEN)-unknown-elf-gcc
-LD=$(PULP_RISCV_GCC_TOOLCHAIN)/bin/$(ISA)$(XLEN)-unknown-elf-gcc
+# Include directories
+INC += -Isw
+INC += -Isw/inc
+INC += -Isw/utils
+
+BOOTSCRIPT := sw/kernel/crt0.S
+LINKSCRIPT := sw/kernel/link.ld
+
+CC=$(ISA)$(XLEN)-unknown-elf-gcc
+LD=$(CC)
 OBJDUMP=$(ISA)$(XLEN)-unknown-elf-objdump
-CC_OPTS=-march=$(ARCH)$(XLEN)$(XTEN) -D__$(ISA)__ -O2 -g -Wextra -Wall -Wno-unused-parameter -Wno-unused-variable -Wno-unused-function -Wundef -fdata-sections -ffunction-sections -MMD -MP
-LD_OPTS=-march=$(ARCH)$(XLEN)$(XTEN) -D__$(ISA)__ -MMD -MP -nostartfiles -nostdlib -Wl,--gc-sections
+CC_OPTS=-march=$(ARCH)$(XLEN)$(XTEN) -mabi=ilp32 -D__$(ISA)__ -O2 -g -Wextra -Wall -Wno-unused-parameter -Wno-unused-variable -Wno-unused-function -Wundef -fdata-sections -ffunction-sections -MMD -MP
+LD_OPTS=-march=$(ARCH)$(XLEN)$(XTEN) -mabi=ilp32 -D__$(ISA)__ -MMD -MP -nostartfiles -nostdlib -Wl,--gc-sections
 
 # Setup build object dirs
 CRT=$(BUILD_DIR)/crt0.o
@@ -68,19 +76,19 @@ VSIM_LIBS=$(BUILD_DIR)/$(TEST_SRCS)/work
 # Build implicit rules
 $(STIM_INSTR) $(STIM_DATA): $(BIN)
 	objcopy --srec-len 1 --output-target=srec $(BIN) $(BIN).s19
-	sw/parse_s19.pl $(BIN).s19 > $(BIN).txt
-	python sw/s19tomem.py $(BIN).txt $(STIM_INSTR) $(STIM_DATA)
+	scripts/parse_s19.pl $(BIN).s19 > $(BIN).txt
+	python scripts/s19tomem.py $(BIN).txt $(STIM_INSTR) $(STIM_DATA)
 	ln -sfn $(INI_PATH) $(VSIM_INI)
 	ln -sfn $(WORK_PATH) $(VSIM_LIBS)
 
-$(BIN): $(CRT) $(OBJ) sw/link.ld
-	$(LD) $(LD_OPTS) -o $(BIN) $(CRT) $(OBJ) -Tsw/link.ld
+$(BIN): $(CRT) $(OBJ)
+	$(LD) $(LD_OPTS) -o $(BIN) $(CRT) $(OBJ) -T$(LINKSCRIPT)
 
-$(CRT): $(BUILD_DIR) sw/crt0.S
-	$(CC) $(CC_OPTS) -c sw/crt0.S -o $(CRT)
+$(CRT): $(BUILD_DIR)
+	$(CC) $(CC_OPTS) -c $(BOOTSCRIPT) -o $(CRT)
 
 $(OBJ): $(TEST_SRCS) $(BUILD_DIR)/$(TEST_SRCS)
-	$(CC) $(CC_OPTS) -c $(TEST_SRCS) $(FLAGS) -Isw -o $(OBJ)
+	$(CC) $(CC_OPTS) -c $(TEST_SRCS) $(FLAGS) $(INC) -o $(OBJ)
 
 $(BUILD_DIR)/$(TEST_SRCS):
 	mkdir -p $(BUILD_DIR)/$(TEST_SRCS)
@@ -102,12 +110,12 @@ ifeq ($(gui), 0)
 	-gSTIM_DATA=stim_data.txt              \
 	-gPROB_STALL=$(P_STALL)
 else
-	cd $(BUILD_DIR)/$(TEST_SRCS);      \
-	$(QUESTA) vsim vopt_tb             \
-	-do "add log -r sim:/redmule_complex_tb/*" \
-	-do "source $(WAVES)"              \
-	-gSTIM_INSTR=stim_instr.txt        \
-	-gSTIM_DATA=stim_data.txt          \
+	cd $(BUILD_DIR)/$(TEST_SRCS); \
+	$(QUESTA) vsim vopt_tb        \
+	-do "add log -r sim:/$(tb)/*" \
+	-do "source $(WAVES)"         \
+	-gSTIM_INSTR=stim_instr.txt   \
+	-gSTIM_DATA=stim_data.txt     \
 	-gPROB_STALL=$(P_STALL)
 endif
 
@@ -121,7 +129,18 @@ bender_defs += -D COREV_ASSERT_OFF
 bender_targs += -t rtl
 bender_targs += -t test
 bender_targs += -t cv32e40p_exclude_tracer
-bender_targs += -t redmule_test
+
+ifeq ($(REDMULE_COMPLEX),1)
+	tb := redmule_complex_tb
+	WAVES := $(mkfile_path)/wave_complex_xif.do
+	bender_targs += -t redmule_complex
+	bender_targs += -t redmule_test_complex
+else
+	tb := redmule_tb
+	WAVES := $(mkfile_path)/wave.do
+	bender_targs += -t redmule_hwpe
+	bender_targs += -t redmule_test_hwpe
+endif
 
 update-ips:
 	$(BENDER) update
@@ -170,7 +189,7 @@ hw-clean-all:
 	rm -rf .cached_ipdb.json
 
 hw-opt:
-	$(QUESTA) vopt +acc=npr -o vopt_tb redmule_complex_tb -floatparameters+redmule_complex_tb -work $(BUILD_DIR)
+	$(QUESTA) vopt +acc=npr -o vopt_tb $(tb) -floatparameters+$(tb) -work $(BUILD_DIR)
 
 hw-compile:
 	$(QUESTA) vsim -c +incdir+$(UVM_HOME) -do 'quit -code [source $(compile_script)]'
