+++ rtl/redmule_wrap.sv
@@ -32,6 +32,8 @@ parameter  int unsigned  DW                 = DATA_W               , // TCDM por
 parameter  int unsigned  MP                 = DW/redmule_pkg::MemDw,
 parameter  type          redmule_data_req_t = logic                ,
 parameter  type          redmule_data_rsp_t = logic                ,
+parameter  type          redmule_ctrl_req_t = logic                ,
+parameter  type          redmule_ctrl_rsp_t = logic                ,
 localparam fp_format_e   FpFormat           = FPFORMAT             , // Data format (default is FP16)
 localparam int unsigned  Height             = ARRAY_HEIGHT         , // Number of PEs within a row
 localparam int unsigned  Width              = ARRAY_WIDTH          , // Number of parallel rows
@@ -72,9 +74,8 @@ localparam int unsigned  BITW               = fp_width(FpFormat)  // Number of b
 
 redmule_data_req_t data_req;
 redmule_data_rsp_t data_rsp;
-
-// hci_core_intf #(.DW(DW)) tcdm (.clk(clk_i));
-hwpe_ctrl_intf_periph #(.ID_WIDTH(ID_WIDTH)) periph (.clk(clk_i));
+redmule_ctrl_req_t ctrl_req;
+redmule_ctrl_rsp_t ctrl_rsp;
 
 generate
   for(genvar ii=0; ii<MP; ii++) begin: tcdm_binding
@@ -91,25 +92,25 @@ generate
   assign data_rsp.r_user  = tcdm_r_user_i;
 endgenerate
 
-always_comb begin
-  periph.req       = periph_req_i;
-  periph.add       = periph_add_i;
-  periph.wen       = periph_wen_i;
-  periph.be        = periph_be_i;
-  periph.data      = periph_data_i;
-  periph.id        = periph_id_i;
-  periph_gnt_o     = periph.gnt;
-  periph_r_data_o  = periph.r_data;
-  periph_r_valid_o = periph.r_valid;
-  periph_r_id_o    = periph.r_id;
-end
+assign ctrl_req.req     = periph_req_i;
+assign ctrl_req.add     = periph_add_i;
+assign ctrl_req.wen     = periph_wen_i;
+assign ctrl_req.be      = periph_be_i;
+assign ctrl_req.data    = periph_data_i;
+assign ctrl_req.id      = periph_id_i;
+assign periph_gnt_o     = ctrl_rsp.gnt;
+assign periph_r_data_o  = ctrl_rsp.r_data;
+assign periph_r_valid_o = ctrl_rsp.r_valid;
+assign periph_r_id_o    = ctrl_rsp.r_id;
 
 redmule_top #(
   .ID_WIDTH           ( ID_WIDTH           ),
   .N_CORES            ( N_CORES            ),
   .DW                 ( DW                 ),
   .redmule_data_req_t ( redmule_data_req_t ),
-  .redmule_data_rsp_t ( redmule_data_rsp_t )
+  .redmule_data_rsp_t ( redmule_data_rsp_t ),
+  .redmule_ctrl_req_t ( redmule_ctrl_req_t ),
+  .redmule_ctrl_rsp_t ( redmule_ctrl_req_t )
 ) i_redmule_top       (
   .clk_i              ( clk_i              ),
   .rst_ni             ( rst_ni             ),
@@ -118,7 +119,8 @@ redmule_top #(
   .busy_o             ( busy_o             ),
   .data_req_o         ( data_req           ),
   .data_rsp_i         ( data_rsp           ),
-  .periph             ( periph             )
+  .ctrl_req_i         ( ctrl_req           ),
+  .ctrl_rsp_o         ( ctrl_rsp           )
 );
 
 endmodule: redmule_wrap
