+++ scripts/s19tomem.py
