+++ sw/hal_redmule.h
@@ -22,7 +22,7 @@
 #ifndef __HAL_REDMULE_H__
 #define __HAL_REDMULE_H__
 
-#include "inc/tensor_dim.h"
+#include "tensor_dim.h"
 
 /*
  *
