+++ sw/redmule_complex.c
@@ -1,0 +1,104 @@
+/*
+ * Copyright (C) 2022-2023 ETH Zurich and University of Bologna
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * SPDX-License-Identifier: Apache-2.0
+ * 
+ * Author: Yvan Tortorella  <yvan.tortorella@unibo.it>
+ *
+ * RedMulE Complex test
+ */
+
+#include <stdint.h>
+#include "redmule_utils.h"
+#include "archi_redmule.h"
+#include "tensor_dim.h"
+#include "x_input.h"
+#include "w_input.h"
+#include "y_input.h"
+#include "z_output.h"
+#include "golden.h"
+
+int main() {
+
+  uint8_t *x = x_inp;
+  uint8_t *w = w_inp;
+  uint8_t *y = y_inp;
+  uint8_t *z = z_oup; // golden_out //1c010000
+
+  uint16_t m_size = M_SIZE;
+  uint16_t n_size = N_SIZE;
+  uint16_t k_size = K_SIZE;
+
+  uint32_t x_addr = *(uint32_t *) &x;
+  uint32_t w_addr = *(uint32_t *) &w;
+  uint32_t y_addr = *(uint32_t *) &y;
+  uint32_t cfg_reg0 = ((k_size << 16) | (m_size << 0));
+  uint32_t cfg_reg1 = (n_size << 0);
+  asm volatile ("addi t0, %0, 0" :: "r"(x_addr));
+  asm volatile ("addi t1, %0, 0" :: "r"(w_addr));
+  asm volatile ("addi t2, %0, 0" :: "r"(y_addr));
+  asm volatile ("addi t3, %0, 0" :: "r"(cfg_reg0));
+  asm volatile ("addi t4, %0, 0" :: "r"(cfg_reg1));
+
+  /* mcfig instruction */
+  // asm volatile(
+  //       ".word (0x0       << 25) | \     /* Empty */
+  //              (0b11101   << 20) | \     /* Rs2 */
+  //              (0b11100   << 15) | \     /* Rs1 */
+  //              (0x00      <<  7) | \     /* Empty */
+  //              (0b0001011 <<  0)   \n"); /* OpCode */
+  
+  asm volatile(
+        ".word (0x0       << 25) | \
+               (0b11101   << 20) | \
+               (0b11100   << 15) | \
+               (0x00      <<  7) | \
+               (0b0001011 <<  0)   \n");
+
+  volatile int errors = 0;
+
+  /* arith instruction */
+  // sm volatile(
+  //      ".word (0b00111   << 27) | \     /* Rs3 */
+  //             (0b00      << 25) | \     /* Empty*/
+  //             (0b00110   << 20) | \     /* Rs2 */
+  //             (0b00101   << 15) | \     /* Rs1 */
+  //             (0b0       << 14) | \     /* Custom format enable/disable */
+  //             (0b0       << 13) | \     /* Widening enable/disable */
+  //             (0b001     << 10) | \     /* Operation selection */
+  //             (0b001     <<  7) | \     /* Data format */
+  //             (0b0101011 <<  0)   \n"); /* OpCode */
+
+  asm volatile(
+        ".word (0b00111   << 27) | \
+               (0b00      << 25) | \
+               (0b00110   << 20) | \
+               (0b00101   << 15) | \
+               (0b0       << 14) | \
+               (0b0       << 13) | \
+               (0b001     << 10) | \
+               (0b001     <<  7) | \
+               (0b0101011 <<  0)   \n");
+
+  // Wait for end of computation
+  asm volatile ("wfi" ::: "memory");
+
+  errors = redmule16_compare_int(y, golden, m_size*k_size/2);
+
+  *(int *) 0x80000000 = errors;
+
+  tfp_printf ("Terminated test with %d errors. See you!\n", errors);
+
+  return errors;
+}
